board: partitions: Set BOARD_NAND_SPARE_SIZE to 0
Not sure what exactly this is supposed to be. I've set the partition
sizes to be the actual number of bytes of the partitions. There is
no extra space after it that could be used. Set this to 0 so
we don't build images that are too big to fit into the partition.