overlay: frameworks/base: Avoid fast brightness fade animation
Since the backlight is implemented using PWM, it will fade
automatically when changing the backlight. Therefore, it's much
more efficient to avoid fading manually and just set the target
brightness directly.

Especially when the display is turned on there can be a certain
delay when changing the brightness value, which may cause the
animation to look weird. Changing to the target brightness directly
fixes it.