audio: Drop parameter framework based Audio HAL configuration
When I started working on the device, it seemed like a good idea
to continue using a similar HAL as used on the stock ROM. However,
the parameter framework based is way to complex and totally
overengineered if you just have speakers, headphones and the
internal microphone.

It also caused some issues that were hard to debug due to its
complexity. The next commit will introduce a much simpler Audio HAL
as a replacement.