media: Remove Intel hardware video codecs and make OMX binderized
Unfortunately, the old OMX implementation for MediaSDK seems to be
broken on Android 9 (Pie), and the new binaries for P no longer
support older chipsets like Bay Trail. Eventually it will be possible
to fix this once the OMX plugins are open-sourced. Until then, remove
them and fall back to software codecs.