Stop registering a WiFi vendor HAL
If there is no WiFi vendor HAL for the chipset, you are not
supposed to register one. WiFi will still work without it.

This means that rfkill is no longer used to turn off WiFi.
However, stopping wpa_supplicant and bringing down the network
interface should have the same effect.