overlay: frameworks/base: Re-enable LiveDisplay/color transformations
When Android introduced its own "night mode", LiveDisplay in LineageOS
was also changed to only enable it on platforms where the color
transformation can be efficiently done in hardware.
At the moment, this ROM does not have a Graphics Composer that is
capable of doing this (and it's not implemented in the kernel either).

Enabling this option makes LiveDisplay working again even without
support in the Graphics Composer, although there will be a
significant overhead for the GPU fallback. Like on Nougat, we will
just need to live with that for now, and leave it up to the user
to enable this feature if they really want to.