Restore hardware accelerated video codecs on Pie, now fully open-source
Closes #54.

It turns out Intel open-sourced the remaining part on
https://github.com/Intel-Media-SDK/MediaSDK_OMX_IL at the end of last
year. I only stumbled on it while randomly browsing on GitHub;
it was not announced anywhere yet.

As expected, it does no longer work as-is. That version is the one
for the newer vaapi driver at https://github.com/intel/media-driver

However, now that it is fully open-source we can patch the
implementation to make it work with the old intel-vaapi-driver.
It took several hours to find what exactly was causing problems,
but eventually I found workarounds to make it decode and encode
most of the videos. (like on Nougat)

I also used the chance to finally fix the encoder, although it does
not perform as good as I hoped - the device still feels quite slow
when using the screen recorder.