wifi: Add service to change WiFi MAC address on boot
wifi-hwaddr opens a netlink socket, and changes the MAC address
of the specified network interface as soon as it is added.

Simply running "ifconfig wlan0 hw <addr>" once at boot does not
work because we don't know exactly when the WiFi driver module has
finished initialization. The service waits until the interface
appears and also handles the case when the WiFi driver is unloaded
and later loaded again.