recovery: root: Symlink /property_contexts to plat_property_contexts
This is needed for installing older ZIPs from Nougat (e.g. the
LineageOS 14.1 builds) because otherwise it cannot read any
properties. This is because the file location changed in Android
Oreo.

Eventually this should be a combination of plat_property_contexts
and vendor_property_contexts, but for now vendor_property_contexts
is empty so there is no need to bother. It's somewhat unlikely
those installation ZIPs really access any vendor specific properties
anyway.