device: Enable Android Go build configuration defaults
At this point, a tablet with 1 GB RAM can be regarded as a "low ram"
device, as modern devices usually come with at least 4 GB RAM.
The Android Go build configuration is specifically optimized for
devices with 1 GB (or even less) RAM. So it makes sense to enable
them for this tablet.

It does come with a price: generally, less RAM is traded for more
CPU usage. Some features like splitscreen get disabled. But all in
all it might improve the situation a bit.