Remove vendor repository, move firmware extraction into this repository
With the hardware-accelerated video codecs open-sourced, there are
no more proprietary user-space libraries used in this ROM.
So it is worth separating the remaining ones (ARM native bridge and
DRM) since those are optional features and it is not clear if they
can actually be redistributed.

Those will be provided separately.

The firmware extraction now resides in a "firmware" directory in this
repository, and documentation was enhanced overall to explain this.